Moving MVVM stuff into a directory
