+++ src/Pickles/Pickles.UserInterface/Pickles.UserInterface.csproj
@@ -73,9 +73,9 @@
     </ApplicationDefinition>
     <Compile Include="ConfigurationValidationService.cs" />
     <Compile Include="Settings\MainModel.cs" />
-    <Compile Include="MultiSelectableCollection.cs" />
-    <Compile Include="SelectableCollection.cs" />
-    <Compile Include="SelectableItem.cs" />
+    <Compile Include="Mvvm\MultiSelectableCollection.cs" />
+    <Compile Include="Mvvm\SelectableCollection.cs" />
+    <Compile Include="Mvvm\SelectableItem.cs" />
     <Compile Include="Settings\MainModelSerializer.cs" />
     <Compile Include="Settings\StreamExtensions.cs" />
     <Page Include="CoolCheckBox.xaml">
@@ -90,9 +90,9 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="EventHandlerExtensions.cs" />
-    <Compile Include="NotifyPropertyChanged.cs" />
-    <Compile Include="RelayCommand.cs" />
+    <Compile Include="Mvvm\EventHandlerExtensions.cs" />
+    <Compile Include="Mvvm\NotifyPropertyChanged.cs" />
+    <Compile Include="Mvvm\RelayCommand.cs" />
     <Compile Include="MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
